requirements.txt
pandas>=2.0.0,<3
PyYAML>=6.0.0,<7
